+ add suckless terminal description from 0.8.4, switching from urxvt
urxvt is nice, but has issues with too much spacing: doesn't allow a
negative lineSpace (and the default with 0 lineSpace for some reason is
much wider than eg xterm lines), and while negative letterSpace works, it
causes many unicode characters (such as arrows) to draw as boxes.  after
much research, basically there is no solution for this, except to switch
to a different terminal.

the suckless 256-color terminal seemed like the best replacement for my
purposes.